requirements.txt
# Core dependencies
requests>=2.31.0
aiohttp>=3.8.0
aiofiles>=23.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
PyYAML>=6.0

# PDF processing
PyPDF2>=3.0.0
pdfplumber>=0.9.0
pytesseract>=0.3.10

# Text analysis and ML
langdetect>=1.0.9
scikit-learn>=1.3.0
nltk>=3.8.1
numpy>=1.24.0
chardet>=5.0.0

# International support
charset-normalizer>=3.0.0

# Database and storage
sqlite3  # Built-in with Python

# Utilities
pathlib  # Built-in with Python
dataclasses  # Built-in with Python
typing  # Built-in with Python
asyncio  # Built-in with Python
multiprocessing  # Built-in with Python

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0